+++ trixnity-messenger-compose-view/src/iosMain/kotlin/de/connect2x/messenger/compose/view/room/settings/ExportRoom.ios.kt
@@ -3,8 +3,25 @@ package de.connect2x.messenger.compose.view.room.settings
 import androidx.compose.runtime.Composable
 import de.connect2x.trixnity.messenger.export.Destination
 import de.connect2x.trixnity.messenger.export.FileBasedExportRoomProperties
+import okio.Path.Companion.toPath
+import platform.Foundation.NSDocumentDirectory
+import platform.Foundation.NSDownloadsDirectory
+import platform.Foundation.NSFileManager
+import platform.Foundation.NSSearchPathForDirectoriesInDomains
+import platform.Foundation.NSUserDomainMask
 
 @Composable
-internal actual fun SelectExportDestination(properties: FileBasedExportRoomProperties?, result: (Destination?) -> Unit) {
-    // TODO
+internal actual fun SelectExportDestination(
+    properties: FileBasedExportRoomProperties?,
+    result: (Destination?) -> Unit
+) {
+    val documentsPath = NSSearchPathForDirectoriesInDomains(NSDocumentDirectory, NSUserDomainMask, true) // TODO
+        .firstOrNull() as? String
+    if (documentsPath == null) {
+        result(null)
+        return
+    }
+
+    println("EEEEEEEEE $documentsPath")
+    result(documentsPath.toPath())
 }