+++ trixnity-messenger/src/iosMain/kotlin/de/connect2x/trixnity/messenger/DI.ios.kt
@@ -1,9 +1,7 @@
 package de.connect2x.trixnity.messenger
 
-import de.connect2x.trixnity.messenger.util.IsNetworkAvailable
 import org.koin.core.module.Module
 import org.koin.dsl.module
 
 actual fun platformModule(): Module = module {
-    single<IsNetworkAvailable> { NetworkMonitor() }
 }