+++ trixnity-messenger/src/iosMain/kotlin/de/connect2x/trixnity/messenger/util/UriCaller.ios.kt
@@ -5,6 +5,8 @@ import org.koin.core.module.Module
 import org.koin.dsl.module
 import platform.Foundation.NSURL.Companion.URLWithString
 import platform.UIKit.UIApplication
+import platform.darwin.dispatch_async
+import platform.darwin.dispatch_get_main_queue
 
 private val log = KotlinLogging.logger { }
 
@@ -14,7 +16,12 @@ actual fun platformUriCallerModule(): Module = module {
             val safeUri = checkNotNull(URLWithString(uri))
             log.info { "call uri: $safeUri" }
             if (!external) log.debug { "does not support internal uri calling yet" }
-            UIApplication.sharedApplication.openURL(safeUri)
+
+            // openURL requires to be called on the main queue, because otherwise it will be called on a background
+            // thread.
+            dispatch_async(dispatch_get_main_queue()) {
+                UIApplication.sharedApplication.openURL(safeUri)
+            }
         }
     }
 }